+++ src/main/kotlin/io/taff/spek/expekt/Config.kt
@@ -1,35 +1,11 @@
 package io.taff.spek.expekt
 
-import com.fasterxml.jackson.databind.DeserializationFeature
-import com.fasterxml.jackson.datatype.jsr310.JavaTimeModule
-import com.fasterxml.jackson.module.kotlin.jacksonObjectMapper
 import mu.NamedKLogging
-import java.time.OffsetDateTime
-import java.time.format.DateTimeFormatter
 
 /**
  * Configuration used throughout this library.
  */
 object Config {
-
-    /**
-     * used for serializing objects for logging as well as deserializing json into lists and maps.
-     */
-    var objectMapper = jacksonObjectMapper().apply {
-        configure(DeserializationFeature.FAIL_ON_UNKNOWN_PROPERTIES, false)
-        registerModule(JavaTimeModule())
-    }
-
-    /**
-     * Used for deserializing dates. When dates are represented as strings (e.g after a partially complete json eserialization),
-     * we use this to attempt deserializing strings to dates for comparison. e.g,assuming today is 2021/10/10:
-     * expect(listOf(OffsetDateTime.now()) {
-     *    containInOrder("2021/10/10")
-     * }
-     */
-    var dateTimeDeserializer: (String) -> OffsetDateTime = { str: String ->
-        OffsetDateTime.parse(str, DateTimeFormatter.ISO_OFFSET_DATE_TIME)
-    }
 
     /**
      * Used for all logging.
@@ -49,6 +25,4 @@ object Config {
 /**
  * Configure this library.
  */
-fun configure(fxn: Config.()-> Unit) {
-    fxn(Config)
-}
+fun configure(fxn: Config.()-> Unit) { fxn(Config) }