+++ src/main/kotlin/io/taff/spek/expekt/Constants.kt
@@ -2,7 +2,11 @@ package io.taff.spek.expekt
 
 import java.math.BigDecimal
 import java.math.BigInteger
+import java.time.Instant
+import java.time.LocalDate
+import java.time.LocalDateTime
 import java.time.OffsetDateTime
+import java.time.ZonedDateTime
 import kotlin.reflect.KClass
 
 /**
@@ -26,7 +30,13 @@ internal val defaultComparers = mutableMapOf<KClass<*>, (Any?, Any?) -> Boolean>
 
     String::class to ::compareComparables,
 
-    Enum::class to ::compareComparables,
+    OffsetDateTime::class to ::compareOffsetDateTimes,
 
-    OffsetDateTime::class to ::compareDateTimes
+    Instant::class to ::compareInstants,
+
+    LocalDate::class to ::compareLocalDates,
+
+    LocalDateTime::class to ::compareLocalDateTimes,
+
+    ZonedDateTime::class to ::compareZonedDateTimes
 )