+++ src/main/kotlin/io/taff/spek/expekt/iterable/Contains.kt
@@ -14,7 +14,7 @@ import io.taff.spek.expekt._compare
  */
 inline fun <T> containInAnyOrder(vararg expecteds: T) = object : Matcher<Iterable<T>> {
 
-    override val description = "containing${Config.objectMapper.writeValueAsString(expecteds)}"
+    override val description = "containInAnyOrder(${expecteds.joinToString { "$it" }})"
 
     override fun invoke(actuals: Iterable<T>) = actuals!!.toList().let { actualList ->
         expecteds.all { expected->
@@ -36,7 +36,7 @@ inline fun <T> containInAnyOrder(vararg expecteds: T) = object : Matcher<Iterabl
         }
     }.let { suceeded ->
         if (suceeded) MatchResult.Match
-        else MatchResult.Mismatch("actual: ${Config.objectMapper.writeValueAsString(actuals)}")
+        else MatchResult.Mismatch("actual: [${actuals.joinToString { "$it" }}]")
     }
 
     override fun toString() = description
@@ -51,7 +51,7 @@ inline fun <T> containInAnyOrder(vararg expecteds: T) = object : Matcher<Iterabl
  */
 inline fun <T> containInOrder(vararg expecteds: T) = object : Matcher<Iterable<T>> {
 
-    override val description = "containingInOrder${Config.objectMapper.writeValueAsString(expecteds)}"
+    override val description = "containInOrder(${expecteds.joinToString { "$it" }})"
 
     override fun invoke(actuals: Iterable<T>) = actuals.toList().let { actualList ->
         expecteds.mapIndexed { index, expected ->
@@ -77,5 +77,5 @@ inline fun <T> containInOrder(vararg expecteds: T) = object : Matcher<Iterable<T
 
     override fun toString() = description
 
-    private fun mismatch(expected: T, actuals: Iterable<T>) =  MatchResult.Mismatch("Actual iterable: ${Config.objectMapper.writeValueAsString(actuals)}\n missing item: $expected")
+    private fun mismatch(expected: T, actuals: Iterable<T>) =  MatchResult.Mismatch("missing item: $expected, actual: [${actuals.joinToString { "$it" }}]")
 }