+++ src/main/kotlin/io/taff/spek/expekt/map/Contains.kt
@@ -11,34 +11,31 @@ import java.lang.IllegalArgumentException
  *
  * values are compared using [io.taff.spek.expekt.Config.comparers]
  * ```
- * mapOf(1 to 2) should beAMapOf(mapOf(1 to 2))
+ * mapOf(1 to 2) should contain(mapOf(1 to 2))
  * ```
  */
-inline fun <K, V> beAMapOf(expectedMap: Map<K, V>) = expectedMap
+inline fun <K, V> contain(expectedMap: Map<K, V>) = expectedMap
     .entries
     .map { it.toPair() }
     .toTypedArray()
-    .let { beAMapOf(*it) }
+    .let { contain(*it) }
 
 /**
  * A matcher that checks whether the expected set of entries are a subset of the actual map.
  *
  * values are compared using [io.taff.spek.expekt.Config.comparers]
  * ```
- * mapOf(1 to 2) shouldNot beAMapOf(2 to 1)
+ * mapOf(1 to 2) shouldNot contain(2 to 1)
  * ```
  */
-inline fun <K, V> beAMapOf(vararg expectedEntries: Pair<K, V>) = object : Matcher<Map<K, V>> {
+inline fun <K, V> contain(vararg expectedEntries: Pair<K, V>) = object : Matcher<Map<K, V>> {
 
-    private val serializedEntries by lazy {
-        """{${
-            expectedEntries.joinToString {
-                "\n\t\"${it.first}\": ${it.second}"
-            }
-        }}""".trimIndent()
+    private val serializedEntries by lazy { expectedEntries
+            .joinToString { "\n\t${it.first} = ${it.second}" }
+            .trimIndent()
     }
 
-    override val description = serializedEntries
+    override val description = "contains($serializedEntries)"
 
     /**
      * Compare the actual map against the expected map.
@@ -51,14 +48,10 @@ inline fun <K, V> beAMapOf(vararg expectedEntries: Pair<K, V>) = object : Matche
         }.let { nonMatchingEntry ->
             if (nonMatchingEntry == null) {
                 MatchResult.Match.also {
-                    Config.logger.info {
-                        "Success! Json object contained attributes {${
-                            expectedEntries.joinToString { "\n\t\"${it.first}\": ${it.second}" }
-                        }\n}"
-                    }
+                    Config.logger.info { "Success! Matcher: $description" }
                 }
             } else {
-                MatchResult.Mismatch("actualMap: $actualMap\n problematic entry: {${nonMatchingEntry.first}: ${nonMatchingEntry.second}}")
+                MatchResult.Mismatch("actual: $actualMap\n problematic entry: {${nonMatchingEntry.first} = ${nonMatchingEntry.second}}")
             }
         }